backend/requirements.txt
#transformers torch
#arabert
#faeasapy
groq
python-dotenv
scikit-learn
matplotlib
seaborn
pandas
numpy
tqdm
loguru
arabic-reshaper
python-bidi